+++ mkr1/LightHTML/ButtonNode.cs
@@ -5,20 +5,17 @@ namespace LightHTML;
 
 class ButtonNode : ElementNode
 {
-    private ICommand? clickCommand;
+    private ICommand? clickCommand, mouseEnterCommand, mouseLeaveCommand;
     public ButtonNode() : base("button", false, false)
     {
 
     }
-    public void SetClickCommand(ICommand? command = null)
-    {
-        clickCommand = command;        
-    }
-    public void PerformClick()
-    {
-        if (clickCommand != null)
-            clickCommand.Execute(this);
-    }
+    public void SetClickCommand(ICommand? command = null) => clickCommand = command;
+    public void SetMouseEnterCommand(ICommand? command = null) => mouseEnterCommand = command;
+    public void SetMouseLeaveCommand(ICommand? command = null) => mouseLeaveCommand = command;
+    public void PerformClick() => clickCommand?.Execute(this);
+    public void PerformMouseEnter() => mouseEnterCommand?.Execute(this);
+    public void PerformMouseLeave() => mouseLeaveCommand?.Execute(this);
 
     public override string OuterHTML
     {
