+++ mkr1/LightHTML/Program.cs
@@ -6,8 +6,8 @@ ElementNode container = new("div");
 
 ButtonNode btn = new();
 container.Children.Add(btn);
-SaveCommand cmd = new();
-btn.SetClickCommand(cmd);
+btn.SetClickCommand(new SaveCommand());
+btn.SetMouseEnterCommand(new HoverCommand());
 
 TextNode txt = new("Click on me!");
 btn.Children.Add(txt);
@@ -16,3 +16,4 @@ string document = container.OuterHTML;
 Console.WriteLine(document);
 
 btn.PerformClick();
+btn.PerformMouseEnter();
